Refactor app config permissions hook: determine permissions from permissions object for CAPI providers, and determine permissions from selfsubjectaccessreviews otherwise
